+++ src/components/snapshots/RepositorySnapshots.vue
@@ -2,7 +2,7 @@
   <q-card>
     <q-card-section class="flex items-center">
       <h1 class="text-h5 q-my-none">{{ t('snapshots.heading', { name: repository }) }}</h1>
-      <reload-button :action="load" />
+      <reload-button :action="load" v-model="snapshotsStore.reloadInterval" />
     </q-card-section>
 
     <q-separator />
@@ -22,9 +22,9 @@ import SnapshotsTable from './SnapshotsTable.vue'
 import { useTranslation } from '../../composables/i18n'
 import { useElasticsearchRequest } from '../../composables/CallElasticsearch'
 import { EsSnapshotList } from '../../composables/components/snapshots/SnapshotRow.ts'
-
+import { useSnapshotsStore } from '../../store/snapshots.ts'
 const t = useTranslation()
-
+const snapshotsStore = useSnapshotsStore()
 const route = useRoute()
 const repository = route.params.repositoryName.toString()