+++ src/components/snapshots/SnapshotsTable.vue
@@ -3,25 +3,25 @@
     <div class="flex items-center">
       <new-snapshot :repository="repository" @reload="emit('reload')" />
       <filter-state
-        v-model="filter"
+        v-model="snapshotsStore.filter"
         :results-count="filterStateProps.resultsCount"
         :filtered-results-count="filterStateProps.filteredResultsCount"
         class="q-ml-md"
       />
     </div>
 
-    <filter-input v-model="filter" />
+    <filter-input v-model="snapshotsStore.filter" />
   </div>
 
   <q-table
     flat
     class="table-mono table-hide-overflow"
     dense
-    row-key="name"
+    row-key="snapshot"
     :columns="columns"
     :rows="filteredResults"
     :rows-per-page-options="DEFAULT_ROWS_PER_PAGE"
-    :pagination="{ sortBy: 'name' }"
+    v-model:pagination="snapshotsStore.pagination"
   >
     <template #body="{ row }">
       <snapshot-row :snapshot="row" :repository="repository" @reload="emit('reload')" />
@@ -36,9 +36,11 @@ import SnapshotRow from './SnapshotRow.vue'
 import NewSnapshot from './NewSnapshot.vue'
 import { SnapshotsTableProps, useSnapshotsTable } from '../../composables/components/snapshots/SnapshotsTable.ts'
 import FilterState from '../shared/FilterState.vue'
+import { useSnapshotsStore } from '../../store/snapshots.ts'
 
+const snapshotsStore = useSnapshotsStore()
 const props = defineProps<SnapshotsTableProps>()
 const emit = defineEmits(['reload'])
 
-const { filter, columns, filteredResults, filterStateProps } = useSnapshotsTable(props)
+const { columns, filteredResults, filterStateProps } = useSnapshotsTable(props)
 </script>